fix bug in grpc services, create db session, add system views

+++ src/server/lsm/disk/page_header.cpp
@@ -14,7 +14,7 @@ Awaitable<void> PageHeader::Flush(sdb::BufferWriter& writer, const PageHeader& h
   co_await writer.WriteInt(header.checksum);
 }
 
-int64_t PageHeader::EstimateSize(const PageHeader& header) {
+size_t PageHeader::EstimateSize(const PageHeader& header) {
   return sdb::BufferWriter::EstimateSize(header.count) + sdb::BufferWriter::EstimateSize(header.checksum);
 }
